Add files via upload
For this Set Exercise,I created a new version of the page that is more accessible for user. I used HTML/CSS to ensure content on the page is accessible visually and via a screen reader,  used HTML to ensure all interactive parts of the page can be accessed using different input devices, used the standard HTML structure and semantic HTML elements shown in class. I can say my code is clearly formatted, indented, and commented.